better link errors
translated over the stuff that was added to the C# impl for better link error messages.
this means switching over from a container of ObjectFile to a container of pairs of string(file name) and ObjectFile.
as the ObjectFile class is getting more complicated (and i plan to add even more to it eventually), even move assignment is getting more expensive.
thus, i've also switched link() over to use std::list instead of std::vector.
we never do tight loops over the entire set of object files (except at the beginning to ensure they're clean), so locality isn't really an issue.