bug fix
fixed a bug where the new TRANS opcode family didn't actually have its handler set in exetables.cpp.
this was causing it to be flagged as an unknown opcode during execution.

found out there was a bug in the 64-bit multiply code because of a bug in the biggerints submodule.
fixed that and merged the updates in.