cpp20 refactor start
switched project over to use some cpp20 features that are now supported on all major compilers.
feature-wise, so far this only makes use of std::endian for supporting big-endian systems.
we used to require little-endian, but now big-endian is supported via byte swapping at runtime.
this makes running on big-endian slower, but is at least supported.
due to std::endian being an enum, it can be in constexpr context, so there is no performance drop for little-endian.

we now require little or big endian at compile time.
distinguishing different types of mixed-endian is not supported by cpp20 std::endian, so neither will we.

changed some of the memory and serialization functions in Utility to support both little/big endian.
some of these have further been made into explicit templates (required type param) to mitigate confusion.
e.g. if the value to write() is a variable, you might think it's a u32, but actually is u64, etc.
additionally, these are now restricted to u8, u16, u32, and u64.

changed memory array used in Computer to std:;vector<u8>.
previously was using manual memory manipulation to achieve 64-byte alignment.
this was done in case i ever decided to use vector intrinsics for speed.
i no longer think that's a good idea, so going with vector and a standard allocator is sufficient and safer.

NULL and EOF are now predefined symbols.
this was done because having to import or redefine them seemed weird from a C perspective.
for NULL, this also makes for faster comparisons because being a pre-define makes it a core constant expression, and so can use special assemble branches like CMPZ.

removed all occurrences of double-underscores from identifiers.
i've since learned that this is undefined behavior for dumb antiquated reasons.

a bunch of other stuff i've forgotten about - not even nearly done with the refactoring.