more refactoring
got rid of a bunch of things that could be done with other things.
for instance, got rid of a few of the private memory access functions.

more mem access is done with safer conventions.
this includes requiring the exact type to be read rather than always returning value zero extended to u64.
apparently 10% of runtime in benchmarking files were from this very thing, so that's neat.

apparently suspended_read wasn't being used by anything other than its access functions.
i've elected to remove it entirely, as i'm not even really sure what it used to do anyway.
it was originally added as a hack fix for blocking console input file read, but is no longer used due to the better way employed now.

began renaming a bunch of things to fit the C++ style.
the name CSX64 comes from "C# x64" i.e. the original impl.
however, this impl has come to be the main one, and deserves proper C++ naming conventions like snake case functions.

found and fixed a bug in one of the vpu handlers thanks to better msvc warnings after an update.

changed Executable to use the new safe serialization functions for its binary io.
this would've caused object files from one endian system to be invalid on others - now is invariantly stored little-endian.

for the longest time the ret value of opcode handlers was discarded by tick() - no longer.
now returning false does not indicate error, rather it's a request for tick() to stop its loop early (if applicable).
so suspended_read is now replaced by sys_read returning false here (and repeating the syscall).

HLT now simply returns false instead of terminating with ErrorCode::Abort.
this is the real-world behavior of HLT, i just didn't find it useful at the time of implementation.
the intent now is that several Computer instances are being tick()ed with a number of ticks in a loop, and this lets it forfeit its time slice early.

changed a bunch of other stuff that i can't even remember... the updated msvc rename/find-uses tools are pretty sick...