UNIT TESTS!!!!
i swore before the saint of the deep that one day i would add unit tests. that day is today.
unit tests are so far very minimalistic, but exist and can be easily expanded.
build support for tests is currently only supported for msvc - it'll be hell to change makemake.py to do test targets.
but i'm not testing any platform-specific logic, so only having unit tests in msvc should be sufficient for the time being.

started changing some register accessors to lowercase, but not completely done.
these commits are considered very experimental...
will eventually be merged, but probably not for quite some time due to slow progress on the refactor.

changed some other stuff too, but not nearly as important ... just renaming and moving stuff around...