more test cases and bug fixes
added some more test cases to cover clock timing (nop) and lea (address computation).
the address computation tests revealed 2 bugs:
1. addressing in n-bit mode was not being truncated to n-bit in the event that it overflowed in 32-bit mode or lower.
2. there were bugs in the register multiplier reducer, causing it to get some multipliers wrong (rare) or false negative fail (common).

both of these are now fixed and the new code was verified with a formal methods and more test cases just to be sure it doesn't break in the future.
the old code was doing all sorts of hard to read things and was causing dereferencing freed memory.
the new code does lock step unique_ptr moves and doesn't reposition anything in the address component path (which was the problem).
also, old code was making copies all over the place (cause C# version did this) - new code only has to create a single copy for distribution of * over +/-, otherwise all are simplified to unique_ptr moves, which are vastly faster.