expr bug fix + times unit tests
added unit tests for times directive.
this lead to finding a bug in the expr evaluator where character literals always evaluated to 0.
this was due to shadowing in the local scope of the character literal handler.

added unit tests for multi-character character literals (in expr context).

changed sstostr() unit testing utility to output like an integer rather than a character (in case of special chars).
in the new character-based unit tests this was causing a problem with debug strings due to inserting null terminators.